refactor: family controller 리팩 완료 #227
