+++ src/main/java/com/ceos/bankids/service/FamilyService.java
@@ -1,10 +1,7 @@
 package com.ceos.bankids.service;
 
 import com.ceos.bankids.domain.Family;
-import com.ceos.bankids.domain.FamilyUser;
 import com.ceos.bankids.domain.User;
-import com.ceos.bankids.dto.KidListDTO;
-import java.util.List;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -12,8 +9,6 @@ public interface FamilyService {
 
     public Family postNewFamily(User user);
     
-    public List<KidListDTO> getKidListFromFamily(FamilyUser familyUser);
-
     public Family getFamilyByCode(String code);
 
     public void deleteFamily(Family family);
